// speedyGo-ui/src/app/services/models/product.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { User } from '../models/user';
export interface Product {
  available?: boolean;
  description?: string;
  discount?: number;
  id?: number;
  imageUrl?: string;
  name?: string;
  partnerProduct?: User;
  price?: number;
  rating?: number;
  stockQuantity?: number;
}